Improve documentation
All of the functions and types are now documented. It isn't perfect but
it should give the reader some sense of what is going on.

Additionally, this solved a potential bug concerning the usage of the
`std::io::Read` trait.

Also some minor changes concerning field types and return types.